Stimulus controller configuration.
