+++ Resources/ChartData/rfs6-packages.json
@@ -1,0 +1,38 @@
+[
+    {
+        "name": "overall",
+        "values": [
+            { "snapshotName": "2024-03-12", "errorCount": 4820 },
+            { "snapshotName": "2024-04-09", "errorCount": 4203 },
+            { "snapshotName": "2024-05-21", "errorCount": 3789 },
+            { "snapshotName": "2024-06-24", "errorCount": 3320 },
+            { "snapshotName": "2024-07-07", "errorCount": 2772 },
+            { "snapshotName": "2024-08-19", "errorCount": 2617 },
+            { "snapshotName": "2024-09-02", "errorCount": 2189 }
+        ]
+    },
+    {
+        "name": "apple",
+        "values": [
+            { "snapshotName": "2024-03-12", "errorCount": 482 },
+            { "snapshotName": "2024-04-09", "errorCount": 420 },
+            { "snapshotName": "2024-05-21", "errorCount": 378 },
+            { "snapshotName": "2024-06-24", "errorCount": 332 },
+            { "snapshotName": "2024-07-07", "errorCount": 277 },
+            { "snapshotName": "2024-08-19", "errorCount": 261 },
+            { "snapshotName": "2024-09-02", "errorCount": 218 }
+        ]
+    },
+    {
+        "name": "sswg",
+        "values": [
+            { "snapshotName": "2024-03-12", "errorCount": 48 },
+            { "snapshotName": "2024-04-09", "errorCount": 42 },
+            { "snapshotName": "2024-05-21", "errorCount": 37 },
+            { "snapshotName": "2024-06-24", "errorCount": 33 },
+            { "snapshotName": "2024-07-07", "errorCount": 27 },
+            { "snapshotName": "2024-08-19", "errorCount": 26 },
+            { "snapshotName": "2024-09-02", "errorCount": 21 }
+        ]
+    }
+]
