--- Resources/Charts/rfs6-errors.json
@@ -1,101 +1,0 @@
-{
-    "$schema": "https://vega.github.io/schema/vega/v5.json",
-    "width": 600,
-    "height": 350,
-    "padding": 5,
-    "data": [
-        {
-            "name": "table",
-            "values": [
-                { "snapshotName": "2024-03-12", "errorCount": 4820 },
-                { "snapshotName": "2024-04-09", "errorCount": 4203 },
-                { "snapshotName": "2024-05-21", "errorCount": 3789 },
-                { "snapshotName": "2024-06-24", "errorCount": 3320 },
-                { "snapshotName": "2024-07-07", "errorCount": 2772 },
-                { "snapshotName": "2024-08-19", "errorCount": 2617 },
-                { "snapshotName": "2024-09-02", "errorCount": 2189 }
-            ]
-        }
-    ],
-    "scales": [
-        {
-            "name": "xscale",
-            "type": "linear",
-            "domain": { "data": "table", "field": "errorCount" },
-            "range": "width",
-            "nice": true
-        },
-        {
-            "name": "yscale",
-            "type": "band",
-            "domain": { "data": "table", "field": "snapshotName" },
-            "range": "height",
-            "padding": 0.2
-        }
-    ],
-    "axes": [
-        {
-            "title": "Total errors across all packages",
-            "orient": "bottom",
-            "scale": "xscale",
-            "labelFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "labelFontSize": { "value": 14 },
-            "titleFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "titleFontSize": { "value": 14 },
-            "titleFontWeight": { "value": "normal" },
-            "titlePadding": { "value": 20 }
-        },
-        {
-            "title": "Swift 6 Snapshot Date",
-            "orient": "left",
-            "scale": "yscale",
-            "labelFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "labelFontSize": { "value": 14 },
-            "titleFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "titleFontSize": { "value": 14 },
-            "titleFontWeight": { "value": "normal" },
-            "titlePadding": { "value": 20 }
-        }
-    ],
-    "marks": [
-        {
-            "type": "rect",
-            "from": { "data": "table" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "value": 0 },
-                    "x2": { "scale": "xscale", "field": "errorCount" },
-                    "y": { "scale": "yscale", "field": "snapshotName" },
-                    "height": { "scale": "yscale", "band": 1 },
-                    "fill": { "value": "#356fce" }
-                }
-            }
-        },
-        {
-            "type": "text",
-            "from": { "data": "table" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "field": "errorCount", "offset": -5 },
-                    "y": { "scale": "yscale", "field": "snapshotName", "offset": { "scale": "yscale", "band": 0.5 } },
-                    "text": { "signal": "datum['errorCount'] + ' errors'" },
-                    "align": { "value": "right" },
-                    "baseline": { "value": "middle" },
-                    "fill": { "value": "white" },
-                    "font": {
-                        "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-                    },
-                    "fontSize": { "value": 14 }
-                }
-            }
-        }
-    ]
-}
