--- Resources/Charts/rfs6-packages.json
@@ -1,145 +1,0 @@
-{
-    "$schema": "https://vega.github.io/schema/vega/v5.json",
-    "width": 600,
-    "height": 350,
-    "padding": 5,
-    "data": [
-        {
-            "name": "overall",
-            "transform": [
-                {
-                    "type": "formula",
-                    "expr": "datetime(datum.snapshot)",
-                    "as": "snapshotDate"
-                }
-            ],
-            "values": [
-                { "snapshot": "2024-03-05", "numPackages": 301 },
-                { "snapshot": "2024-03-18", "numPackages": 329 },
-                { "snapshot": "2024-03-26", "numPackages": 360 },
-                { "snapshot": "2024-04-15", "numPackages": 561 },
-                { "snapshot": "2024-05-09", "numPackages": 819 },
-                { "snapshot": "2024-06-30", "numPackages": 1200 },
-                { "snapshot": "2024-07-29", "numPackages": 1926 }
-            ]
-        },
-        {
-            "name": "apple-packages",
-            "transform": [
-                {
-                    "type": "formula",
-                    "expr": "datetime(datum.snapshot)",
-                    "as": "snapshotDate"
-                }
-            ],
-            "values": [
-                { "snapshot": "2024-03-05", "numPackages": 30 },
-                { "snapshot": "2024-03-18", "numPackages": 42 },
-                { "snapshot": "2024-03-26", "numPackages": 43 },
-                { "snapshot": "2024-04-15", "numPackages": 65 },
-                { "snapshot": "2024-05-09", "numPackages": 89 },
-                { "snapshot": "2024-06-30", "numPackages": 89 },
-                { "snapshot": "2024-07-29", "numPackages": 122 }
-            ]
-        }
-    ],
-    "scales": [
-        {
-            "name": "xscale",
-            "type": "time",
-            "domain": [{ "signal": "datetime('2024-03-01')" }, { "signal": "datetime('2024-12-31')" }],
-            "range": "width"
-        },
-        {
-            "name": "yscale",
-            "type": "linear",
-            "domain": { "data": "overall", "field": "numPackages" },
-            "range": "height",
-            "nice": true
-        }
-    ],
-    "axes": [
-        {
-            "orient": "bottom",
-            "scale": "xscale",
-            "grid": true,
-            "labelFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "labelFontSize": { "value": 14 },
-            "labelAngle": { "value": -45 },
-            "labelAlign": { "value": "right" }
-        },
-        {
-            "orient": "left",
-            "scale": "yscale",
-            "labelFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "labelFontSize": { "value": 14 },
-            "title": "Number of compatible packages",
-            "titlePadding": { "value": 20 },
-            "titleFont": {
-                "value": "-apple-system, BlinkMacSystemFont, 'SF Hello', 'Segoe UI', Roboto, 'Helvetica Neue', Arial, sans-serif"
-            },
-            "titleFontWeight": { "value": "normal" },
-            "titleFontSize": { "value": 14 }
-        }
-    ],
-    "marks": [
-        {
-            "type": "line",
-            "from": { "data": "overall" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "field": "snapshotDate" },
-                    "y": { "scale": "yscale", "field": "numPackages" },
-                    "stroke": { "value": "#356fce" },
-                    "strokeWidth": { "value": 3 }
-                }
-            }
-        },
-        {
-            "type": "symbol",
-            "from": { "data": "overall" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "field": "snapshotDate" },
-                    "y": { "scale": "yscale", "field": "numPackages" },
-                    "size": { "value": 60 },
-                    "fill": { "value": "#356fce" },
-                    "tooltip": {
-                        "signal": "timeFormat(datum.snapshotDate, '%b %d, %Y') + ' - ' + datum.numPackages + ' packages'"
-                    }
-                }
-            }
-        },
-        {
-            "type": "line",
-            "from": { "data": "apple-packages" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "field": "snapshotDate" },
-                    "y": { "scale": "yscale", "field": "numPackages" },
-                    "stroke": { "value": "#7ff01a" },
-                    "strokeWidth": { "value": 3 }
-                }
-            }
-        },
-        {
-            "type": "symbol",
-            "from": { "data": "apple-packages" },
-            "encode": {
-                "enter": {
-                    "x": { "scale": "xscale", "field": "snapshotDate" },
-                    "y": { "scale": "yscale", "field": "numPackages" },
-                    "fill": { "value": "#7ff01a" },
-                    "size": { "value": 60 },
-                    "tooltip": {
-                        "signal": "timeFormat(datum.snapshotDate, '%b %d, %Y') + ' - ' + datum.numPackages + ' packages'"
-                    }
-                }
-            }
-        }
-    ]
-}
