+++ Sources/App/Views/ReadyForSwift6/ReadyForSwift6Show+Model.swift
@@ -12,9 +12,35 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
+import Foundation
 import Plot
 
 enum ReadyForSwift6Show {
     struct Model {
+        func readyForSwift6Chart(identifier: String) -> Node<HTML.BodyContext> {
+            let chartDataPath = Current.fileManager.workingDirectory().appending("Resources/ChartData/\(identifier).json")
+            guard let chartData = Current.fileManager.contents(atPath: chartDataPath)?.compactJson()
+            else { return .p("Couldn’t load chart data.") }
+
+            return .div(
+                .class("vega-chart"),
+                .data(named: "controller", value: "vega-chart"),
+                .script(
+                    .data(named: "vega-chart-target", value: "data"),
+                    .attribute(named: "type", value: "application/json"),
+                    .raw(chartData)
+                )
+            )
+        }
+    }
+}
+
+private extension Data {
+    func compactJson() -> String? {
+        guard let json = try? JSONSerialization.jsonObject(with: self),
+              let compactedJsonData = try? JSONSerialization.data(withJSONObject: json),
+              let compactJson = String(data: compactedJsonData, encoding: .utf8)
+        else { return nil }
+        return compactJson
     }
 }
