+++ package.json
@@ -11,7 +11,8 @@
         "@hotwired/stimulus": "^3.2.1",
         "@hotwired/turbo": "^7.3.0",
         "highlight.js": "^11.9.0",
-        "normalize.css": "^8.0.1"
+        "normalize.css": "^8.0.1",
+        "vega": "^5.28.0"
     },
     "devDependencies": {
         "esbuild": "^0.20.2",
